# Optimize tracing enablement logic for evaluation