--- a/tests/langchain/test_langchain_evaluation.py
+++ b/tests/langchain/test_langchain_evaluation.py
@@ -12,6 +12,7 @@ from mlflow.models.evaluation import evaluate
 from mlflow.models.evaluation.evaluators.default import DefaultEvaluator
 from mlflow.tracing.constant import TraceMetadataKey
 
+from tests.openai.test_openai_evaluate import purge_traces
 from tests.tracing.helper import get_traces, reset_autolog_state  # noqa: F401
 
 _EVAL_DATA = pd.DataFrame(
@@ -37,7 +38,7 @@ def create_fake_chain():
 
 
 @pytest.mark.parametrize(
-    "original_autolog_config",
+    "config",
     [
         None,
         {"log_traces": False},
@@ -46,9 +47,13 @@ def create_fake_chain():
     ],
 )
 @pytest.mark.usefixtures("reset_autolog_state")
-def test_langchain_evaluate(original_autolog_config):
-    if original_autolog_config:
-        mlflow.langchain.autolog(**original_autolog_config)
+def test_langchain_evaluate(config):
+    if config:
+        mlflow.langchain.autolog(**config)
+        mlflow.openai.autolog(**config)  # Our chain contains OpenAI call as well
+
+    is_trace_disabled = config and not config.get("log_traces", True)
+    is_trace_enabled = config and config.get("log_traces", True)
 
     chain = create_fake_chain()
 
@@ -66,22 +71,25 @@ def test_langchain_evaluate(original_autolog_config):
             )
         log_model_mock.assert_not_called()
 
-    assert len(get_traces()) == 2
-    for trace in get_traces():
-        assert len(trace.data.spans) == 5
-    assert run.info.run_id == get_traces()[0].info.request_metadata[TraceMetadataKey.SOURCE_RUN]
+    # Traces should not be logged when disabled explicitly
+    if is_trace_disabled:
+        assert len(get_traces()) == 0
+    else:
+        assert len(get_traces()) == 2
+        for trace in get_traces():
+            assert len(trace.data.spans) == 5
+        assert run.info.run_id == get_traces()[0].info.request_metadata[TraceMetadataKey.SOURCE_RUN]
+
+    purge_traces()
 
     # Test original langchain autolog configs is restored
     with mock.patch("mlflow.langchain.log_model") as log_model_mock:
         chain.invoke({"question": "text"})
 
-        if original_autolog_config and original_autolog_config.get("log_models", False):
+        if config and config.get("log_models", False):
             log_model_mock.assert_called_once()
 
-        if original_autolog_config and original_autolog_config.get("log_traces", True):
-            assert len(get_traces()) == 3
-        else:
-            assert len(get_traces()) == 2
+        assert len(get_traces()) == (1 if is_trace_enabled else 0)
 
 
 @pytest.mark.usefixtures("reset_autolog_state")