--- a/tests/llama_index/test_llama_index_evaluate.py
+++ b/tests/llama_index/test_llama_index_evaluate.py
@@ -7,6 +7,7 @@ import mlflow.utils.autologging_utils
 from mlflow.metrics import latency
 from mlflow.tracing.constant import TraceMetadataKey
 
+from tests.openai.test_openai_evaluate import purge_traces
 from tests.tracing.helper import get_traces, reset_autolog_state  # noqa: F401
 
 _EVAL_DATA = pd.DataFrame(
@@ -24,7 +25,7 @@ _EVAL_DATA = pd.DataFrame(
 
 
 @pytest.mark.parametrize(
-    "original_autolog_config",
+    "config",
     [
         None,
         {"log_traces": False},
@@ -32,9 +33,13 @@ _EVAL_DATA = pd.DataFrame(
     ],
 )
 @pytest.mark.usefixtures("reset_autolog_state")
-def test_llama_index_evaluate(single_index, original_autolog_config):
-    if original_autolog_config:
-        mlflow.llama_index.autolog(**original_autolog_config)
+def test_llama_index_evaluate(single_index, config):
+    if config:
+        mlflow.llama_index.autolog(**config)
+        mlflow.openai.autolog(**config)  # Our model contains OpenAI call as well
+
+    is_trace_disabled = config and not config.get("log_traces", True)
+    is_trace_enabled = config and config.get("log_traces", True)
 
     engine = single_index.as_query_engine()
 
@@ -50,16 +55,18 @@ def test_llama_index_evaluate(single_index, original_autolog_config):
         )
     assert eval_result.metrics["latency/mean"] > 0
 
-    # Traces should be automatically enabled during evaluation
-    assert len(get_traces()) == 2
-    assert run.info.run_id == get_traces()[0].info.request_metadata[TraceMetadataKey.SOURCE_RUN]
+    # Traces should not be logged when disabled explicitly
+    if is_trace_disabled:
+        assert len(get_traces()) == 0
+    else:
+        assert len(get_traces()) == 2
+        assert run.info.run_id == get_traces()[0].info.request_metadata[TraceMetadataKey.SOURCE_RUN]
+
+    purge_traces()
 
     # Test original autolog configs is restored
     engine.query("text")
-    if original_autolog_config and original_autolog_config.get("log_traces", True):
-        assert len(get_traces()) == 3
-    else:
-        assert len(get_traces()) == 2
+    assert len(get_traces()) == (1 if is_trace_enabled else 0)
 
 
 @pytest.mark.parametrize("engine_type", ["query", "chat"])